seaice3p/params/convection.py
from dataclasses import dataclass
from serde import serde, coerce
from .dimensional import DimensionalParams, DimensionalRJW14Params, NoBrineConvection


@serde(type_check=coerce)
@dataclass(frozen=True)
class RJW14Params:
    """Parameters for the RJW14 parameterisation of brine convection"""

    Rayleigh_salt: float = 44105
    Rayleigh_critical: float = 2.9
    convection_strength: float = 0.13
    couple_bubble_to_horizontal_flow: bool = False
    couple_bubble_to_vertical_flow: bool = False
    advective_heat_flux_in_ocean: bool = True


BrineConvectionParams = RJW14Params | NoBrineConvection


def get_dimensionless_brine_convection_params(
    dimensional_params: DimensionalParams,
) -> BrineConvectionParams:
    match dimensional_params.brine_convection_params:
        case DimensionalRJW14Params():
            return RJW14Params(
                Rayleigh_salt=dimensional_params.Rayleigh_salt,
                Rayleigh_critical=dimensional_params.brine_convection_params.Rayleigh_critical,
                convection_strength=dimensional_params.brine_convection_params.convection_strength,
                couple_bubble_to_horizontal_flow=dimensional_params.brine_convection_params.couple_bubble_to_horizontal_flow,
                couple_bubble_to_vertical_flow=dimensional_params.brine_convection_params.couple_bubble_to_vertical_flow,
                advective_heat_flux_in_ocean=dimensional_params.brine_convection_params.advective_heat_flux_in_ocean,
            )
        case NoBrineConvection():
            return NoBrineConvection()
